+++ src/App.jsx
@@ -4,18 +4,18 @@ import { useState } from 'react';
 import './styles/reset.css';
 import './styles/common.css';
 import Content from './router/Content';
-import { light, dark } from './styles/theme_mode';
+import { light, dark } from './styles/theme';
 
 function App() {
   const [isDarkMode, setIsDarkMode] = useState(false);
-  const theme_mode = isDarkMode ? dark : light;
+  const theme = isDarkMode ? dark : light;
 
   const toggleTheme = () => {
     setIsDarkMode((prevMode) => !prevMode);
   };
 
   return (
-    <ThemeProvider theme={theme_mode}>
+    <ThemeProvider theme={theme}>
       <button onClick={toggleTheme}>
         Switch to {isDarkMode ? 'Light' : 'Dark'} Mode
       </button>
