--- src/styles/theme_mode.js
@@ -1,27 +1,0 @@
-export const light = {
-  background: 'var(--white)', // 메인 배경색
-  surface: 'var(--surface)', // 카드 또는 섹션 배경색
-  blackText: 'var(--black)', // 검정 텍스트
-  whiteText: 'var(--white)', // 흰색 텍스트
-  text: 'var(--gray-900)', // 기본 텍스트 색상
-  headerName: 'var(--gray-800)', // 헤더 또는 주요 텍스트 색상
-  textarea: 'var(--gray-600)', // 입력 필드 등 배경색
-  secondary: 'var(--gray-500)', // 부가적인 설명 텍스트 색상
-  dateText: 'var(--gray-400)', // 날짜 또는 메타 정보 텍스트
-  border: 'var(--gray-300)', // 경계선 색상
-  line: 'var(--gray-200)', // 구분선 또는 미세한 경계 색상
-};
-
-export const dark = {
-  background: 'var(--gray-900)', // 메인 배경색
-  surface: 'var(--gray-800)', // 카드 또는 섹션 배경색
-  blackText: 'var(--white)', // 검정 텍스트 대신 흰색 텍스트로 변경
-  whiteText: 'var(--black)', // 흰색 텍스트에 대응하는 색상 (검정)
-  text: 'var(--gray-100)', // 기본 텍스트 색상
-  headerName: 'var(--gray-200)', // 헤더 또는 주요 텍스트 색상
-  textarea: 'var(--gray-300)', // 입력 필드 등 배경색
-  secondary: 'var(--gray-400)', // 부가적인 설명 텍스트 색상
-  dateText: 'var(--gray-500)', // 날짜 또는 메타 정보 텍스트
-  border: 'var(--gray-600)', // 경계선 색상
-  line: 'var(--gray-700)', // 구분선 또는 미세한 경계 색상
-};
